refactor: enhance avatar and user profile components
- Updated Avatar component to explicitly import React for clarity.
- Refactored EnhancedAvatar component by removing unused state and props, simplifying its structure.
- Added `forceDecoration` prop to EnhancedAvatar and UserProfile components to control avatar decorations.
- Introduced admin controls for avatar decoration testing in ProfilePageClient, allowing admins to test different avatar styles.
- Added corresponding styles for admin controls in profile-page.scss to improve UI consistency.

These changes improve code clarity, maintainability, and enhance the user experience for avatar customization.
