+++ src/app/components/avatar/Avatar.tsx
@@ -3,7 +3,7 @@
 import { adventurer } from '@dicebear/collection';
 import { createAvatar } from '@dicebear/core';
 import { useAtom } from 'jotai';
-import { memo, useCallback, useEffect, useMemo, useState } from 'react';
+import React, { memo, useCallback, useEffect, useMemo, useState } from 'react';
 import { avatarCacheAtom } from '../../../lib/state/atoms';
 import { InteractiveTooltip } from '../tooltip/InteractiveTooltip';
 import './Avatar.css';
